requirements.txt
Flask==3.0.3
APScheduler==3.10.4
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==22.0.0
python-dotenv==1.0.1
